Revert "添加了一些功能，并优化了代码结构"
